backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
psycopg2==2.9.7
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-decouple==3.8
django-extensions==3.2.3
reportlab==4.0.4
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.0
